add docs + remove comment + fix devices
