+++ types/puppeteer/Input.d.ts
@@ -1,3 +1,8 @@
+/**
+ * Definition from Input.js
+ * https://github.com/GoogleChrome/puppeteer/blob/master/lib/Input.js
+ */
+
 export type KeyboardKeyF = 'F1' | 'F2' | 'F3' | 'F4' | 'F5' | 'F6' | 'F7' | 'F8' | 'F9' | 'F10' | 'F11' | 'F12' | 'F13' | 'F14' | 'F15' | 'F16' | 'F17' | 'F18' | 'F19' | 'F20' | 'F21' | 'F22' |
  'F23' | 'F24';
 export type KeyboardKeyDigits = '0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | 'Digit0' | 'Digit1' | 'Digit2' | 'Digit3' | 'Digit4' | 'Digit5' | 'Digit6' | 'Digit7' | 'Digit8' |
