+++ types/puppeteer/index.d.ts
@@ -11,21 +11,34 @@
 
 /// <reference types="node" />
 
-export * from './Accessibility';
-export * from './common';
 export * from './able';
+export * from './Accessibility';
 export * from './Browser';
 export * from './BrowserFetcher';
+export * from './Connection';
 export * from './Coverage';
+export * from './DeviceDescriptors';
 export * from './Dialog';
+// export * from './EmulationManager'; // no public interface
 export * from './Errors';
+// export * from './Events'; // no public interface
 export * from './ExecutionContext';
 export * from './FrameManager';
 export * from './Input';
 export * from './JSHandle';
 export * from './Launcher';
+// export * from './LifecycleWatcher'; // no public interface
+// export * from './Multimap'; // no public interface
 export * from './NetworkManager';
 export * from './Page';
+// export * from './PipeTransport'; // no public interface
 export * from './Puppeteer';
 export * from './Target';
+// export * from './TaskQueue'; // no public interface
+// export * from './TimeoutSettings'; // no public interface
+// export * from './Tracing'; // no public interface
+// export * from './USKeyboardLayout'; // no public interface
+// export * from './WebSocketTransport'; // no public interface
 export * from './Worker';
+// export * from './api'; // no public interface
+// export * from './helper'; // no public interface
